requirements.txt
fastapi==0.115.12
httpx==0.28.1
pydantic==2.11.3
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.40
uvicorn==0.34.2
pytest==8.3.5
pytest-cov==6.1.1